added test_script to each folder
test_script runs level curves in every parameter and generates random samples inside a given variation around a set of base parameters. Percent error is calculated and plotted using either an exact solution or dielectric function solution as the "true" solution. The idea is to run this and feel confident in the accuracy of the methods, then run active subspace.